Cocci: LKMP: Generate type test suite
These scripts are for generating the test suite, which
will be used in the LKMP Coccinelle project for Spring 2020.

gen_c.cocci generates a C file consisting of a sequence of
declarations for each type found in one or more of the input files.

gen_cocci.cocci generates the semantic patch to match against
the C file. The generated semantic patch consists of multiple
rules, each to match against one declaration. A match is considered
valid if the line it is found on is between on the start or end of
the declaration in the C file. A list of failed rules is output
after completion.

The Makefile can be used to interact with the scripts.

To generate the scripts:
	make FILES="<filenames_separated_by_space>"
or:
	make DIR=<directory_name>

To run the test suite:
	make test

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>